Change groupData storage, account for timestamp in old key display
Refactor to only store a list of groupData objects to localStorage.
Other representations created on init, and maintained throughout.

Keep track of groups based on owner+name, and calculate when groupData
for that group changed.
We can compare message timestamps (taken from message ID) against this
to know if a stale key was used.